requirements.txt
streamlit
pandas
requests
matplotlib
python-telegram-bot
python-dotenv